+++ NEWS.md
@@ -1,6 +1,8 @@
 # GenoBrew 0.0.4
 
 * Add box to edit specific samples CNV estimations and replace/drop in the main dataset
+* Add tutorial link
+* Add shinyapps.io github action deployment
 
 # GenoBrew 0.0.3